+++ gulpfile.babel.js
@@ -1,22 +1,34 @@
-'use strict'; 'use strong';
+import {default as register} from   'babel-core/register';
+register({
+    only: [
+        '**/node_modules/angie*/**',
+        '**/src/**'
+    ],
+    stage: 0
+});
 
 // System Modules
-import gulp from                'gulp';
-import {exec} from              'child_process';
-import eslint from              'gulp-eslint';
-import jscs from                'gulp-jscs';
-import istanbul from            'gulp-istanbul';
-import {Instrumenter} from      'isparta';
-import mocha from               'gulp-mocha';
-import cobertura from           'istanbul-cobertura-badger';
+import fs from                      'fs';
+import gulp from                    'gulp';
+import {argv} from                  'yargs';
+import {exec} from                  'child_process';
+import eslint from                  'gulp-eslint';
+import jscs from                    'gulp-jscs';
+import istanbul from                'gulp-istanbul';
+import {Instrumenter} from          'isparta';
+import mocha from                   'gulp-mocha';
+import cobertura from               'istanbul-cobertura-badger';
+import babel from                   'gulp-babel';
+import {bold, red} from             'chalk';
 
-const src = 'src/**/*.js',
-      testSrc = 'test/**/*.spec.js',
-      docSrc = 'doc',
-      coverageDir = 'coverage';
+const SRC = 'src/**/*.js',
+    TRANSPILED_SRC = 'dist',
+    TEST_SRC = 'test/**/*.spec.js',
+    DOC_SRC = 'doc',
+    COVERAGE_SRC = 'coverage';
 
 gulp.task('eslint', function () {
-    gulp.src([ src, testSrc ]).pipe(
+    gulp.src([ SRC, TEST_SRC ]).pipe(
         eslint()
     ).pipe(
         eslint.format()
@@ -25,7 +37,7 @@ gulp.task('eslint', function () {
     );
 });
 gulp.task('jscs', [ 'eslint' ], function () {
-    return gulp.src([ src, 'test/**/!(decorators)*.spec.js' ])
+    return gulp.src([ SRC, 'test/**/!(decorators)*.spec.js' ])
         .pipe(jscs({
             fix: true,
             configPath: '.jscsrc',
@@ -34,8 +46,8 @@ gulp.task('jscs', [ 'eslint' ], function () {
 });
 gulp.task('mocha', function(cb) {
     let proc;
-    new Promise(function(resolve, reject) {
-        proc = gulp.src(src).pipe(
+    new Promise(function(resolve) {
+        proc = gulp.src(SRC).pipe(
             istanbul({
                 instrumenter: Instrumenter,
                 includeUntested: true
@@ -49,7 +61,7 @@ gulp.task('mocha', function(cb) {
             ).pipe(mocha({
                 reporter: 'spec'
             }).on('error', function(e) {
-                resolve();
+                throw new Error(e);
             }).on('end', function() {
                 resolve();
             }));
@@ -66,14 +78,31 @@ gulp.task('mocha', function(cb) {
         return cobertura('coverage/cobertura-coverage.xml', 'svg', cb);
     });
 });
+gulp.task('babel', function() {
+    return gulp.src(SRC).pipe(babel()).pipe(gulp.dest(TRANSPILED_SRC));
+});
 gulp.task('esdoc', function(cb) {
     exec('esdoc -c esdoc.json', cb);
 });
+gulp.task('bump', function(cb) {
+    const version = argv.version,
+        bump = (f) => fs.writeFileSync(f, fs.readFileSync(f, 'utf8').replace(
+            /[0-9]{1,2}\.[0-9]{1,2}\.[0-9]{1,2}/,
+            version
+        ));
+    if (version) {
+        bump('bin/angie');
+        bump('bin/angie-dist');
+        bump('package.json');
+    } else {
+        throw new Error(bold(red('No version specified!!')));
+    }
+});
 gulp.task('watch', [ 'jscs', 'mocha' ], function() {
-    gulp.watch([ src, testSrc, '../gh-pages-angie/**' ], [ 'mocha' ]);
+    gulp.watch([ SRC, TEST_SRC ], [ 'mocha' ]);
 });
 gulp.task('watch:mocha', [ 'jscs', 'mocha' ], function() {
-    gulp.watch([ src, testSrc, '../gh-pages-angie/**' ], [ 'mocha' ]);
+    gulp.watch([ SRC, TEST_SRC ], [ 'mocha' ]);
 });
 gulp.task('test', [ 'jscs', 'mocha' ]);
-gulp.task('default', [ 'jscs', 'mocha' ]);
+gulp.task('default', [ 'jscs', 'mocha', 'babel', 'esdoc' ]);