+++ src/util/scaffold/project.js
@@ -1,24 +1,45 @@
-'use strict'; 'use strong';
+/**
+ * @module project.js
+ * @author Joe Groseclose <@benderTheCrime>
+ * @date 8/16/2015
+ */
 
 // System Modules
+// Do not alias this as the commands mirror the global `confirm` and `prompt`
+import {default as promptly} from       'promptly';
 import fs from                          'fs';
 import util from                        'util';
+import chalk, {
+    bold,
+    green
+} from                                  'chalk';
 import $LogProvider from                'angie-log';
 
 // Angie Modules
 import {$StringUtil} from                '../Util';
 
-const p = process;
+const p = process,
+      breen = (v) => bold(green(v));
 
 /**
  * @desc $$createProject is the function called when the CLI attempts to create
- * a project from the command line. Scaffolds the main folder in the specified
- * location or no folder and then folders for commonly used providers and the
- * Angie config file, AngieFile.json.
+ * a project from the command line. This scaffolds the main folder in the
+ * specified location or no folder and then folders for commonly used providers
+ * and the Angie config file (AngieFile.json).
+ *
+ * The CLI function to create a project will ask the user a series of questions.
+ * The result of these questions will be passed to the AngieFile.json:
+ *     - By default the `development` option is set to true
+ *     - By default, the `databases` object has one sqlite3 database, but this
+ * database is not instantiated
+ *     - User will be prompted for caching of static assets
+ *     - User will be prompted for default script file attachment
  *
  * This function will gracefully exit the process if successful and exit with
  * errors if unsuccessful.
  * @since 0.0.1
+ * @todo If any more CL arguments are required, you must abstract confirm/prompt
+ * as functions
  * @param {object} args A list of arugments passed from the CLI parser
  * @param {string} args.name The name of the project being created. This must
  * consist of letters, dashes, & underscores
@@ -54,8 +75,6 @@ export default function $$createProject(args = {}) {
     mkDirFiles = mkDir ? `${mkDir}/` : '';
     mkSub = `${mkDirFiles}src`.replace(/\/{2}/g, '/');
 
-    console.log('MKDIR', mkDir);
-
     try {
 
         // We cannot create a dir if the argument is empty
@@ -77,6 +96,7 @@ export default function $$createProject(args = {}) {
 
         // Create static folders
         [
+            'test',
             'static',
             'templates'
         ].forEach(function(v) {
@@ -86,29 +106,88 @@ export default function $$createProject(args = {}) {
         throw new $$ProjectCreationError(e);
     } finally {
 
-        // Read our AngieFile template and reproduce in the target directory
-        let template = fs.readFileSync(
-            `${__dirname}/../../templates/AngieFile.template.json`,
-            'utf8'
-        );
-        template = util.format(template, name, name);
-        fs.writeFileSync(
-            `${mkDirFiles}AngieFile.json`,
-            template,
-            'utf8'
-        );
-    }
+        // This is where we create our AngieFile, we can pick certain values with
+        // which we can populate our config:
 
-    $LogProvider.info('Project successfully created');
-    p.exit(0);
+        // cacheStaticAssets
+        let staticCache = false,
+
+            // Default JS to be loaded with all HTML files
+            defaultAppJavaScriptFilename;
+
+        // Wrap the prompts in a Promise
+        new Promise(function(resolve) {
+            promptly.confirm(
+                `${breen('Do you want Angie to cache static assets?')} :`,
+                resolve
+            );
+        }).then(function(v) {
+            staticCache = !!v;
+            return;
+        }).then(function() {
+
+            // Ask what the default JS filename should be
+            return new Promise(function(resolve) {
+                return promptly.prompt(
+                    `${breen(
+                        'What would you like to call the "default" ' +
+                        'loaded script file ' +
+                        `(${bold(chalk.white('default is'))} ` +
+                        `${chalk.cyan('application.js')})?`
+                     )} :`,
+                     {
+                         default: 'application.js',
+                         validator: function(v) {
+                             if (v && v.indexOf('.js') === -1) {
+                                 throw new Error(
+                                     bold(chalk.red(
+                                         'Input must be a valid ".js" file.'
+                                     ))
+                                 );
+                             }
+                             return v.replace(/\/|\\/g, '');
+                         }
+                     },
+                    function(e, v) {
+                        resolve(v);
+                    }
+                );
+            });
+        }).then(function(v) {
+            defaultAppJavaScriptFilename = v;
+        }).then(function() {
+
+            // Read our AngieFile template and reproduce in the target directory
+            let template = fs.readFileSync(
+                `${__dirname}/../../templates/json/AngieFile.template.json`,
+                'utf8'
+            );
+            template = util.format(
+                template,
+                name,
+                name,
+                staticCache,
+                defaultAppJavaScriptFilename
+            );
+            fs.writeFileSync(
+                `${mkDirFiles}AngieFile.json`,
+                template,
+                'utf8'
+            );
+            fs.writeFileSync(
+                `${mkDirFiles}static/${defaultAppJavaScriptFilename}`,
+                ''
+            );
+
+            $LogProvider.info('Project successfully created');
+            p.exit(0);
+        });
+    }
 }
 
 class $$ProjectCreationError extends Error {
     constructor(e) {
         $LogProvider.error(e);
         super(e);
-        p.exit(1);
     }
 }
-
-// TODO add test folder, src folder inside