+++ test/src/Server.spec.js
@@ -1,0 +1,17 @@
+// // Test Modules
+// import {assert, expect} from        'chai';
+// import simple, {mock, spy} from     'simple-mock';
+//
+// // System Modules
+// import repl from                    'repl';
+//
+//
+// // Angie Modules
+// import app from                      '../../src/Angie';
+// import {$$watch, $$shell} from       '../../src/Server';
+//
+// const P = process;
+//
+// describe('$$watch', function() {});
+// describe('$$shell', function() {});
+// describe('$$server', function() {});